requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
email-validator==2.0.0.post2
python-dotenv==1.0.0
gunicorn==20.1.0
